nightvision/requirements.txt
astral
pause
RPi.GPIO
